+++ package.json
@@ -19,9 +19,12 @@
     "eslint": "8.33.0",
     "eslint-config-next": "13.1.6",
     "framer-motion": "^9.0.1",
+    "joi": "^17.8.3",
+    "jotai": "^2.0.2",
     "next": "13.1.6",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "typescript": "4.9.5"
+    "typescript": "4.9.5",
+    "use-between": "^1.3.5"
   }
 }
